- used 2 space tab -  used static_cast instead of c style cast - noexcept correctness - std::bind_front in threadpool

+++ Minecraft Clone/src/Level/Entity/Type/FallingBlock.cpp
@@ -4,55 +4,54 @@
 #include "Level/Entity/Entity.h"
 
 void FallingBlock::Tick(Entity* entity, Dimension* dimension) {
-    //    std::cout << entity->Properties.Velocity.y << "\n";
-    float mspt = 1.0f / static_cast<float>(dimension->tick_rate_);
+  //    std::cout << entity->Properties.Velocity.y << "\n";
+  float mspt = 1.0f / static_cast<float>(dimension->tick_rate_);
 
-    // Physics
+  // Physics
 
-    // Logger.LogInfo("Sand", std::to_string(entity->Properties.Position.y));
+  // Logger.LogInfo("Sand", std::to_string(entity->Properties.Position.y));
 
-    entity->properties_.acceleration_.y =
-        -dimension->world_->parameters.gravity_;
+  entity->properties_.acceleration_.y = -dimension->world_->parameters.gravity_;
 
-    entity->properties_.velocity_ += entity->properties_.acceleration_ * mspt;
+  entity->properties_.velocity_ += entity->properties_.acceleration_ * mspt;
 
-    int distanceCheck = (int)ceil(abs(entity->properties_.velocity_.y * mspt));
+  int distanceCheck =
+      static_cast<int>(ceil(abs(entity->properties_.velocity_.y * mspt)));
 
-    float collusionDistance =
-        dimension->collusion_detector_->TraceSingleAxisCollision(
-            entity->properties_.position_, Directions<BlockPos>::kDown,
-            distanceCheck + 1);
+  float collusionDistance =
+      dimension->collusion_detector_->TraceSingleAxisCollision(
+          entity->properties_.position_, Directions<BlockPos>::kDown,
+          distanceCheck + 1);
 
-    float timeTillCollusion =
-        abs(collusionDistance / entity->properties_.velocity_.y);
+  float timeTillCollusion =
+      abs(collusionDistance / entity->properties_.velocity_.y);
 
-    bool collideWithGround = false;
+  bool collideWithGround = false;
 
-    if ((timeTillCollusion < mspt) && (collusionDistance != -1.f)) {
-        entity->properties_.position_[1] +=
-            entity->properties_.velocity_[1] * (float)timeTillCollusion;
-        entity->properties_.velocity_[1] = 0.f;
-        entity->properties_.acceleration_[1] = 0.f;
+  if ((timeTillCollusion < mspt) && (collusionDistance != -1.f)) {
+    entity->properties_.position_[1] +=
+        entity->properties_.velocity_[1] * timeTillCollusion;
+    entity->properties_.velocity_[1] = 0.f;
+    entity->properties_.acceleration_[1] = 0.f;
 
-        collideWithGround = true;
-    } else {
-        entity->properties_.position_[1] +=
-            entity->properties_.velocity_[1] * mspt;
-    }
-    entity->is_dirty_ = true;
+    collideWithGround = true;
+  } else {
+    entity->properties_.position_[1] += entity->properties_.velocity_[1] * mspt;
+  }
+  entity->is_dirty_ = true;
 
-    if (collideWithGround) {
-        BlockEvent addBlock{BlockPos{(int)entity->properties_.position_.x,
-                                     (int)entity->properties_.position_.y,
-                                     (int)entity->properties_.position_.z},
-                            g_blocks.SAND, g_event_handler.BlockPlace};
-        dimension->event_manager_.AddEvent(addBlock);
+  if (collideWithGround) {
+    BlockEvent addBlock{BlockPos{entity->properties_.position_.x,
+                                 entity->properties_.position_.y,
+                                 entity->properties_.position_.z},
+                        g_blocks.SAND, g_event_handler.BlockPlace};
+    dimension->event_manager_.AddEvent(addBlock);
 
-        EntityEvent removeEntity;
-        removeEntity.id_ = g_event_handler.RemoveEntity;
-        removeEntity.entity_uuid_ = entity->properties_.entity_uuid_;
-        removeEntity.unique_id_ = 50;
+    EntityEvent removeEntity;
+    removeEntity.id_ = g_event_handler.RemoveEntity;
+    removeEntity.entity_uuid_ = entity->properties_.entity_uuid_;
+    removeEntity.unique_id_ = 50;
 
-        dimension->event_manager_.AddEvent(removeEntity);
-    }
+    dimension->event_manager_.AddEvent(removeEntity);
+  }
 }
