+++ Minecraft Clone/src/Level/TerrainGeneration/GenLayer/GenLayerFuzzyZoom.h
@@ -4,13 +4,12 @@
 #include "Level/TerrainGeneration/GenLayer/IntCache.h"
 
 class GenLayerFuzzyZoom : public GenLayerZoom {
-   public:
-    GenLayerFuzzyZoom(long long baseSeedIn, GenLayer* parentIn)
-        : GenLayerZoom(baseSeedIn, parentIn) {}
+ public:
+  GenLayerFuzzyZoom(int64_t baseSeedIn, GenLayer* parentIn)
+      : GenLayerZoom(baseSeedIn, parentIn) {}
 
-    int selectModeOrRandom(int p_151617_1_, int p_151617_2_, int p_151617_3_,
-                           int p_151617_4_) {
-        return selectRandom(
-            {p_151617_1_, p_151617_2_, p_151617_3_, p_151617_4_});
-    }
+  int selectModeOrRandom(int p_151617_1_, int p_151617_2_, int p_151617_3_,
+                         int p_151617_4_) {
+    return selectRandom({p_151617_1_, p_151617_2_, p_151617_3_, p_151617_4_});
+  }
 };
